refactor datagrid schema code, add datetime type detection
